Minor fixes and QoL improvements.
